ADD _verify_project() and _verify_ticket_id() to all tools
Added _verify_project() and _verify_ticket_id() methods to all tools.
These methods are called when a new Ticket object is initiated to verify
that the project and optional ticket_id parameters are valid.

Also calling _verify_ticket_id() in set_ticket_id() in ticket.py to
verify that the new ticket_id is valid.