ADD 'Return Statements' section to README
- Added section to README describing the new return statements in
ticketutil.

Note: In the example code, I assigned the Ticket object to a variable
named 'ticket' and the return statements of each method to a variable
named 't'. For example:

ticket = JiraTicket(. . . .)
t = ticket.create(. . . .)
print(t.status)

This syntax is more in line with other python documentation out there.
I did not make this change in the docs/ and examples/ folders because it
should be made in the new readthedocs documentation instead.