Added new kwargs parameter to handle additional fields passed to resolve the ticket status.
This commit satifies CONTENT-2123.

With the future state of JIRA no project will automatically set resolution when changing the status of the ticket to RESOLVED. RH support confirmed this when a support of the same was created. We will be handling the additional fields required to resolve the tickets via automation in TicketUtil itself.

Modified the change_status method by adding a kwargs parameter that will handle the additional fields required to either set the resolution or a add a comment while resolving a ticket or both so that JIRA tickets can be moved to RESOLVED state with TicketUtil automation.