Retry also HTTP POST requests [RHELDST-31841]
The code already has retry logic that retries HTTP requests in case
of server failures or when the client is rate-limited (HTTP error 429).
However, by default, the retry logic only works for GET requests,
not POST requests. This commit adds POST to the list of methods
allowed to retry requests, in addition to the methods retried
by the requests module.